Add function returning home cage antenna
In experiments with two chambers -- home chamber and stimulus chamber --
there are just two antennas: one at each entrance to a pipe between the home
chamber and stimuls chamber. At the beginning of the experiment mice are
placed in the home cage antenna. The first antenna reading happend when
the first mouse leaves the home chamber and enters the pipe.
Using this function only makes sense for 2 chamber experiments.