Lower necessary number of mouse readouts
Mice are treated as legitimate experiment subjects, if they have been
recorded sufficient number of times and are present for sufficient
number of days. There are some quirks of the hardware/software that
sometimes generate so called "ghost-tags" and also experimentalist
sometimes test the setup by placing a transporder next to an antenna.