Change computation of anticipated following count
Calculate anticipated following of mouse 1 by mouse 2 count as
a mean number of crossings of a chosen tunnel by mouse 2 in a phase
divided to intervals of the length of total time spent by mouse 1
in the tunnel during the phase.

This value is extremely sensitive to overall mice activity and also
to any errors in recording of mouse position (especially situations,
when a mouse spends a lot of time in one of the tunnels or the system
fails to record the mouse leaving the tunnel).