Assume that time is registered in UTC
EcoHAB registeres naive time provided by the comupter time clock. For
the sake of Nencki lab assume that this clock is set to UTC (or any
time zone that does not observe daylight savings time). This needs to be
fixed and in the future user should provide timezones if necessary or,
alternatively, timezone should be save in registrations or time of
registrations should be saved in seconds to epoch or whatever else reasonable
solution we can find.